mass storage and effects working

+++ src/mpc/DACControl.h
@@ -9,7 +9,7 @@
 #define	DACCONTROL_H
 
 void dac_setup();
-void dac_put(int value);
+void dac_put(float value);
 
 #endif	/* DACCONTROL_H */
